src/main/webapp/app/entities/abc-6/delete/abc-6-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Abc6Service } from '../service/abc-6.service';

import { Abc6DeleteDialogComponent } from './abc-6-delete-dialog.component';

describe('Abc6 Management Delete Component', () => {
  let comp: Abc6DeleteDialogComponent;
  let fixture: ComponentFixture<Abc6DeleteDialogComponent>;
  let service: Abc6Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [Abc6DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Abc6DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Abc6DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Abc6Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
